+++ src/main/java/com/starkcore/Settings.java
@@ -5,4 +5,5 @@ public abstract class Settings {
     public static User user = null;
     public static String userAgentOverride = null;
     public static String language = "en-US";
+    public static Integer timeout = null;
 }